fix typos

+++ ratelimiting_kern.c
@@ -25,7 +25,7 @@
 /* Stores the ratelimit value(per second) */
 struct bpf_map_def SEC("maps") rl_config_map = {
 	.type		= BPF_MAP_TYPE_ARRAY,
-	.key_size	= sizeof(uint64_t),
+	.key_size	= sizeof(int),
 	.value_size	= sizeof(uint64_t),
 	.max_entries	= 1,
 };
